backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
pillow
python-multipart